Extract DatastoreWrapperTestEnv.
